+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.1][1.2.1] - 2020-05-15
+
+### Added
+
+- New Looker IP for GCP/Singapore Region
+
+### Fixed
+
+- `gcp_ips` output was incorrect, didn't have GCP IPs.
+
 ## [1.2.0][1.2.0] - 2020-04-1
 
 ### Added
@@ -41,9 +51,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Initial Release
 
-[unreleased]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.2.0...HEAD
-[1.1.2]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.2.0...1.1.2
-[1.1.2]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.1.2...1.1.1
-[1.1.1]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.1.1...1.1.0
-[1.1.0]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.1.0...1.0.0
+[unreleased]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.2.1...HEAD
+[1.2.1]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.2.0...1.2.1
+[1.2.0]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.1.2...1.2.0
+[1.1.2]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.1.1...1.1.1
+[1.1.2]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.1.1...1.1.2
+[1.1.1]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.1.0...1.1.1
+[1.1.0]: https://github.com/captn3m0/terraform-data-looker-ips/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/captn3m0/terraform-data-looker-ips/releases/tag/1.0.0